Pull more tests into main BillingWrapper, create helper verification method
