+++ feature/google/src/testUnityIAP/java/BillingWrapperBC4Test.kt
@@ -151,181 +151,6 @@ class BillingWrapperBC4Test {
     }
 
     @Test
-    fun canMakeAPurchase() {
-        every {
-            mockClient.launchBillingFlow(any(), any())
-        } returns billingClientOKResult
-
-        val skuDetails = stubSkuDetails(productId = "product_a")
-
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-        wrapper.makePurchaseAsync(
-            mockActivity,
-            appUserId,
-            skuDetails.toStoreProduct(),
-            mockReplaceSkuInfo(),
-            "offering_a"
-        )
-
-        verify {
-            mockClient.launchBillingFlow(
-                eq(mockActivity),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun properlySetsBillingFlowParams() {
-        mockkStatic(BillingFlowParams::class)
-        mockkStatic(BillingFlowParams.SubscriptionUpdateParams::class)
-
-        val mockBuilder = mockk<BillingFlowParams.Builder>(relaxed = true)
-        every {
-            BillingFlowParams.newBuilder()
-        } returns mockBuilder
-
-        val skuDetailsSlot = slot<SkuDetails>()
-        every {
-            mockBuilder.setSkuDetails(capture(skuDetailsSlot))
-        } returns mockBuilder
-
-        val mockSubscriptionUpdateParamsBuilder =
-            mockk<BillingFlowParams.SubscriptionUpdateParams.Builder>(relaxed = true)
-        every {
-            BillingFlowParams.SubscriptionUpdateParams.newBuilder()
-        } returns mockSubscriptionUpdateParamsBuilder
-
-        val oldSkuPurchaseTokenSlot = slot<String>()
-        every {
-            mockSubscriptionUpdateParamsBuilder.setOldSkuPurchaseToken(capture(oldSkuPurchaseTokenSlot))
-        } returns mockSubscriptionUpdateParamsBuilder
-
-        val prorationModeSlot = slot<Int>()
-        every {
-            mockSubscriptionUpdateParamsBuilder.setReplaceSkusProrationMode(capture(prorationModeSlot))
-        } returns mockSubscriptionUpdateParamsBuilder
-
-        val sku = "product_a"
-        @BillingClient.SkuType val skuType = BillingClient.SkuType.SUBS
-
-        val upgradeInfo = mockReplaceSkuInfo()
-        val skuDetails = stubSkuDetails(productId = sku, type = skuType)
-
-        val slot = slot<BillingFlowParams>()
-        every {
-            mockClient.launchBillingFlow(eq(mockActivity), capture(slot))
-        } answers {
-            val capturedSkuDetails = skuDetailsSlot.captured
-
-            assertThat(sku).isEqualTo(capturedSkuDetails.sku)
-            assertThat(skuType).isEqualTo(capturedSkuDetails.type)
-
-            assertThat(upgradeInfo.oldPurchase.purchaseToken).isEqualTo(oldSkuPurchaseTokenSlot.captured)
-            assertThat(upgradeInfo.prorationMode).isEqualTo(prorationModeSlot.captured)
-            billingClientOKResult
-        }
-
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-        wrapper.makePurchaseAsync(
-            mockActivity,
-            appUserId,
-            skuDetails.toStoreProduct(),
-            upgradeInfo,
-            null
-        )
-    }
-
-    @Test
-    fun `obfuscatedAccountId is set for non-transfer purchases`() {
-        val mockBuilder = setUpForObfuscatedAccountIDTests()
-
-        wrapper.makePurchaseAsync(
-            mockActivity,
-            appUserId,
-            stubSkuDetails(productId = "product_a").toStoreProduct(),
-            null,
-            null
-        )
-
-        val expectedUserId = appUserId.sha256()
-        verify {
-            mockBuilder.setObfuscatedAccountId(expectedUserId)
-        }
-
-        clearStaticMockk(BillingFlowParams::class)
-    }
-
-    @Test
-    fun `obfuscatedAccountId is not set for transfer purchases`() {
-        val mockBuilder = setUpForObfuscatedAccountIDTests()
-
-        wrapper.makePurchaseAsync(
-            mockActivity,
-            appUserId,
-            stubSkuDetails(productId = "product_a").toStoreProduct(),
-            mockReplaceSkuInfo(),
-            null
-        )
-
-        verify(exactly = 0) {
-            mockBuilder.setObfuscatedAccountId(any())
-        }
-
-        clearStaticMockk(BillingFlowParams::class)
-    }
-
-    @Test
-    fun whenSkuDetailsIsNullPassAnEmptyListToTheListener() {
-        mockNullSkuDetailsResponse()
-
-        val productIDs = setOf("product_a")
-
-        var receivedList: List<StoreProduct>? = null
-        wrapper.querySkuDetailsAsync(
-            ProductType.SUBS,
-            productIDs, {
-                receivedList = it
-            }, {
-                fail("shouldn't be an error")
-            })
-        wrapper.onBillingSetupFinished(billingClientOKResult)
-        assertThat(receivedList).isNotNull
-        assertThat(receivedList!!.size).isZero()
-    }
-
-    @Test
-    fun `getting all purchases gets both subs and inapps`() {
-        val billingClientPurchaseHistoryListenerSlot = slot<PurchaseHistoryResponseListener>()
-        every {
-            mockClient.queryPurchaseHistoryAsync(
-                any(),
-                capture(billingClientPurchaseHistoryListenerSlot)
-            )
-        } answers {
-            billingClientPurchaseHistoryListenerSlot.captured.onPurchaseHistoryResponse(
-                billingClientOKResult,
-                listOf(stubPurchaseHistoryRecord())
-            )
-        }
-
-        var receivedPurchases = listOf<StoreTransaction>()
-        wrapper.queryAllPurchases("appUserID", {
-            receivedPurchases = it
-        }, { fail("Shouldn't be error") })
-
-        assertThat(receivedPurchases.size).isNotZero()
-
-        verify(exactly = 1) {
-            mockClient.queryPurchaseHistoryAsync(BillingClient.SkuType.SUBS, any())
-        }
-
-        verify(exactly = 1) {
-            mockClient.queryPurchaseHistoryAsync(BillingClient.SkuType.INAPP, any())
-        }
-    }
-
-    @Test
     fun `when querying anything and billing client returns an empty list, returns an empty list`() {
         mockQueryPurchasesAsyncResult(null, billingClientOKResult, listOf())
