+++ test-utils/build.gradle
@@ -6,6 +6,8 @@ android {
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation project(":common")
+    implementation project(":public")
+    implementation project(":feature:google")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     bc5SupportDevImplementation "com.android.billingclient:billing:$billing5Version"
     latestDependenciesImplementation "com.android.billingclient:billing:$billing4Version"
